% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{insertFuncTempAddin}
\alias{insertFuncTempAddin}
\title{Insert function template}
\usage{
insertFuncTempAddin()
}
\description{
Call this function as an addin to insert \code{ function(x) \{x\} } at the cursor position.
}
